Move volume fader into its own entity
