+++ entities.cpp
@@ -9,6 +9,77 @@
 #include "entities.h"
 #include "globals.h"
 
+namespace Volume
+{
+	Fader::Fader(QMediaPlayer *player,const QString &arguments,QObject *parent) : Fader(player,0,static_cast<std::chrono::seconds>(0),parent)
+	{
+		Parse(arguments);
+	}
+
+	Fader::Fader(QMediaPlayer *player,unsigned int volume,std::chrono::seconds duration,QObject *parent) : QObject(parent),
+		settingDefaultDuration(SETTINGS_CATEGORY_VOLUME,"DefaultFadeSeconds",5),
+		player(player),
+		initialVolume(static_cast<unsigned int>(player->volume())),
+		targetVolume(std::clamp<int>(volume,0,100)),
+		startTime(TimeConvert::Now()),
+		duration(TimeConvert::Milliseconds(duration))
+	{
+		connect(this,&Fader::AdjustmentNeeded,this,&Fader::Adjust,Qt::QueuedConnection);
+	}
+
+	void Fader::Parse(const QString &text)
+	{
+		QStringList values=text.split(" ",StringConvert::Split::Behavior(StringConvert::Split::Behaviors::SKIP_EMPTY_PARTS));
+		if (values.size() < 1) throw std::range_error("No volume specified");
+		targetVolume=StringConvert::PositiveInteger(values[0]);
+		duration=values.size() > 1 ? static_cast<std::chrono::seconds>(StringConvert::PositiveInteger(values[1])) : settingDefaultDuration;
+	}
+
+	void Fader::Start()
+	{
+		emit Print(QString("Adjusting volume from %1% to %2% over %3 seconds").arg(
+			StringConvert::Integer(initialVolume),
+			StringConvert::Integer(targetVolume),
+			StringConvert::Integer(TimeConvert::Seconds(duration).count())
+		));
+		emit AdjustmentNeeded();
+	}
+
+	void Fader::Stop()
+	{
+		disconnect(this,&Fader::AdjustmentNeeded,this,&Fader::Adjust);
+		deleteLater();
+	}
+
+	void Fader::Abort()
+	{
+		Stop();
+		player->setVolume(initialVolume); // FIXME: volume doesn't return to initial value
+	}
+
+	double Fader::Step(const double &secondsPassed)
+	{
+		// uses x^4 method described here: https://www.dr-lex.be/info-stuff/volumecontrols.html#ideal3
+		// no need to find an ideal curve, this is a bot, not a DAW
+		double totalSeconds=static_cast<double>(duration.count())/TimeConvert::Milliseconds(TimeConvert::OneSecond()).count();
+		int volumeDifference=targetVolume-initialVolume;
+		return std::pow(secondsPassed/totalSeconds,4)*(volumeDifference)+initialVolume;
+	}
+
+	void Fader::Adjust()
+	{
+		double secondsPassed=static_cast<double>((TimeConvert::Now()-startTime).count())/TimeConvert::Milliseconds(TimeConvert::OneSecond()).count();
+		if (secondsPassed > TimeConvert::Seconds(duration).count())
+		{
+			deleteLater();
+			return;
+		}
+		double adjustment=Step(secondsPassed);
+		player->setVolume(adjustment);
+		emit AdjustmentNeeded();
+	}
+}
+
 namespace File
 {
 	List::List(const QString &path)
